Move types into matching files
